Simplify rand generator

+++ internal/config/config.go
@@ -34,8 +34,8 @@ var Config = struct {
 	RPSLimit              int      `env:"RPS_LIMIT" envDefault:"10"`
 	ConnectionsLimit      int      `env:"CONNECTIONS_LIMIT" envDefault:"50"`
 	MaxBodySize           int64    `env:"MAX_BODY_SIZE" envDefault:"10485760"`      // 10 MB
-	SSEMaxLifetime        int      `env:"SSE_MAX_LIFETIME" envDefault:"7200"`       // 2 hours in seconds
-	SSEMaxLifetimeJitter  int      `env:"SSE_MAX_LIFETIME_JITTER" envDefault:"900"` // up to 15 min in seconds
+	SSEMaxLifetime        int64    `env:"SSE_MAX_LIFETIME" envDefault:"7200"`       // 2 hours in seconds
+	SSEMaxLifetimeJitter  int64    `env:"SSE_MAX_LIFETIME_JITTER" envDefault:"900"` // up to 15 min in seconds
 	RateLimitsByPassToken []string `env:"RATE_LIMITS_BY_PASS_TOKEN"`
 
 	// Security
